fix(Graph): 🪲️Minor chores

+++ src/include/engine/id_generator.h
@@ -4,8 +4,8 @@
 #include <cstdint>
 
 struct IdGenerator {
-  uint64_t id = 0;
-  uint64_t generate_id() { return ++id; }
+  uint32_t id = 0;
+  uint32_t generate_id() { return ++id; }
 };
 
 #endif //NAG_ID_GENERATOR_H
